requirements.txt
nba_api==1.5.2
xgboost==2.1.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
boto3==1.34.144
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
pyarrow==16.1.0
pdfplumber==0.11.3
tqdm==4.66.4
openai==1.35.13
